Extract MockPagingIterable for reuse in other tests
